requirements.txt
langchain
langchain_community
python-dotenv
gradio
langchain_google_genai
ipykernel
langchain-groq
langchain-cohere